requirements.txt
boto3==1.9.151
botocore==1.12.151
gunicorn==20.0.4
Flask==1.1.1
gevent==1.4.0
h5py==2.10.0
image==1.5.27
imageio==2.5.0
jsonschema==2.6.0
Keras==2.4.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.1.0
numpy==1.16.1
opencv-python==4.1.0.25
Pillow==6.0.0
tensorflow==2.2.0
Werkzeug==1.0.1
google-cloud-storage==1.31.2
scikit-image==0.17.2